requirements.txt
Flask
flask-bcrypt
flask-jwt-extended
pymongo
pywebpush
python-dotenv